third_party/blink/web_tests/html/select/select-typeahead-non-latin-expected.txt

Check that type-to-select in unopened popups works with Hebrew.

ש
ד
ג
כ
ע
SUCCESS
